back/requirements.txt
fastapi
uvicorn
python-multipart
stegano
pillow